pb部分字段类型进行了调整，从double降到float的

+++ mqttlibrary/src/main/java/us/nonda/mqttlibrary/model/GPSBean.kt
@@ -3,8 +3,8 @@ package us.nonda.mqttlibrary.model
 class GPSBean {
     var lat: Double? = 0.0
     var lng: Double? = 0.0
-    var spd: Double? = 0.0
-    var acc: Double? = 0.0
-    var brg: Double? = 0.0
+    var spd: Float? = 0f
+    var acc: Float? = 0f
+    var brg: Float? = 0f
     var time: Long? = 0
 }
